Fixed SSS with new diffusion profile system

+++ com.unity.render-pipelines.high-definition/Runtime/Material/SubsurfaceScattering/ShaderVariablesSubsurfaceScattering.cs
@@ -13,7 +13,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public fixed float _TransmissionTintsAndFresnel0[16 * 4];  // RGB = 1/4 * color, A = fresnel0
         [HLSLArray(16, typeof(Vector4))]
         public fixed float _WorldScales[16 * 4];        // X = meters per world unit; Y = world units per meter
-        [HLSLArray(16, typeof(uint))]
+        [HLSLArray(16, typeof(float))]
         public fixed uint _DiffusionProfileHashTable[16]; // TODO: constant
 
         // Warning: Unity is not able to losslessly transfer integers larger than 2^24 to the shader system.
